requirements.txt
PyMuPDF==1.24.5
openai==1.33.0
slack_sdk==3.28.0
PyYAML==6.0.1
langchain==0.2.3
langchain-chroma
langchain-openai==0.1.8